Tests_Selenium_page_object/requirements.txt
allure-pytest==2.7.0
pytest==5.4.1
pytest-xdist==1.31.0
selenium==3.141.0
webdriver-manager==2.3.0
